+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.0 - 2021-06-01
+
+### Changed
+
+- Renamed `Volley.LinearSubscription` to `Volley.InOrderSubscription`
+
+### Fixed
+
+- Fixed subscriptions to empty streams in `Volley.InOrderSubscription`.
+
 ## 0.1.3 - 2021-05-03
 
 ### Changed